requirements.txt
# Core Flask application
Flask>=2.0.0

# Keyboard and mouse automation
pyautogui>=0.9.53
PyGetWindow>=0.0.9

# Clipboard support for non-QWERTY keyboards
pyperclip>=1.8.0

# System tray support
pystray>=0.19.0
Pillow>=9.0.0

# HTTP requests for update checking
requests>=2.28.0
